# Use FxHash for faster entry hashing